finished the mobile view for holes
