+++ app/src/components/tournaments/TournamentHoles.tsx
@@ -1,35 +1,43 @@
 import React, { useEffect } from 'react';
+import { holeResultResource } from '../../services/DTOs';
 
 const TournamentHoles = (props: { round: any }) => {
-  console.log(props.round);
-
-  const frontNine = [1, 2, 3, 4, 5, 6, 7, 8, 9];
-  const backNine = [10, 11, 12, 13, 14, 15, 16, 17, 18];
-
-  let frontNineResults: number[] = [];
-  let backNineResults: number[] = [];
-  useEffect(() => {
-    console.log(props.round.holeresults);
-    frontNineResults = props.round.holeresults.splice(0, 8);
-    console.log(frontNineResults);
-    backNineResults = props.round.holeresults.splice(-8);
-    console.log(backNineResults);
-  }, [props]);
+  const generateHoleResultColor = (hole: holeResultResource): string => {
+    if (hole.score === 1) {
+      return 'bg-ace';
+    }
+    const diff = hole.score - hole.par;
+    switch (diff) {
+      case -2:
+        return 'bg-eagle';
+      case -1:
+        return 'bg-birdie';
+      case 0:
+        return '';
+      case 1:
+        return 'bg-bogie';
+      case 2:
+        return 'bg-doubleBogie';
+      case 3:
+        return 'bg-tripleBogie';
+    }
+    return '';
+  };
 
   return (
     <div>
-      <div className="flex flex-col">
-        <div>Hole:</div>
-        <div>Par:</div>
-        <div>Score:</div>
-      </div>
-      <div className="grid grid-cols-9 grid-rows-3">
-        {frontNine.map(number => (
-          <div>{number}</div>
-        ))}
-        {frontNineResults.map((hole: any) => (
-          <div>{hole.score}</div>
-        ))}
+      <div className="flex flex-row">
+        <div className="flex flex-col">
+          <div>F9:</div>
+          <div>B9:</div>
+        </div>
+        <div className="grid w-full grid-cols-9 grid-rows-2 text-center">
+          {props.round.holeResults.map((hole: holeResultResource) => (
+            <div className={' ' + generateHoleResultColor(hole)}>
+              {hole.score}
+            </div>
+          ))}
+        </div>
       </div>
     </div>
   );
