+++ app/tailwind.config.cjs
@@ -12,7 +12,13 @@ module.exports = {
         dgsecondary: '#15616D',
         dgtertiary: '#FF7D00',
         dgbackground: '#001524',
-        dgsoftwhite: '#D9D9D9'
+        dgsoftwhite: '#D9D9D9',
+        ace: '#93c947',
+        eagle: '#557fb5',
+        birdie: '#38659f',
+        bogie: '#c4876a',
+        doubleBogie: '#c86b42',
+        tripleBogie: '#b14d22',
       },
       fontFamily: {
         main: ['Montserrat']
